静默用户(fix):修改查询问题

+++ src/main/java/com/xiliulou/electricity/service/impl/UserActiveInfoServiceImpl.java
@@ -166,7 +166,7 @@ public class UserActiveInfoServiceImpl implements UserActiveInfoService {
     @DS("slave_1")
     public R queryList(UserActiveInfoQuery query) {
         //默认为30天
-        int day = Objects.isNull(query.getDay()) ? 30 : query.getDay();
+        long day = Objects.isNull(query.getDay()) ? 30 : query.getDay();
         query.setLimitTime(System.currentTimeMillis() - day * 24 * 3600000);
         
         List<UserActiveInfoVo> userActiveInfoList = userActiveInfoMapper.queryList(query);
@@ -186,7 +186,7 @@ public class UserActiveInfoServiceImpl implements UserActiveInfoService {
     @Override
     @DS("slave_1")
     public R queryCount(UserActiveInfoQuery query) {
-        int day = Objects.isNull(query.getDay()) ? 30 : query.getDay();
+        long day = Objects.isNull(query.getDay()) ? 30 : query.getDay();
         query.setLimitTime(System.currentTimeMillis() - day * 24 * 3600000);
         
         Long count = userActiveInfoMapper.queryCount(query);
